+++ CHANGELOG.md
@@ -34,6 +34,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Renamed `ChangeDate` rule to `ChangeOnDateRule`.
 - Renamed `ChangeInterval` rule to `ChangeWithIntervalRule`.
 - Renamed `CharacterClass` rule to `CharacterClassRule`.
+- Renamed `Dictionary` rule to `DictionaryRule`.
 
 ### Fixed
 - Nothing